refactor out common tests
