+++ tests/unit/async_db/V1/conftest.py
@@ -1,12 +1,10 @@
-from datetime import date, datetime
 from json import loads
 from re import Pattern, compile
-from typing import Dict
 
 import httpx
 from pytest import fixture
 
-from firebolt.async_db import ARRAY, DECIMAL, Connection, Cursor, connect
+from firebolt.async_db import Connection, Cursor, connect
 from firebolt.client import AsyncClientV1 as Client
 from firebolt.client.auth.base import Auth
 from firebolt.client.auth.username_password import UsernamePassword
@@ -156,31 +154,3 @@ def check_credentials_callback(user: str, password: str, access_token: str) -> C
         )
 
     return check_credentials
-
-
-@fixture
-def types_map() -> Dict[str, type]:
-    base_types = {
-        "int": int,
-        "long": int,
-        "float": float,
-        "double": float,
-        "text": str,
-        "date": date,
-        "pgdate": date,
-        "timestamp": datetime,
-        "timestampntz": datetime,
-        "timestamptz": datetime,
-        "Nothing null": str,
-        "Decimal(123, 4)": DECIMAL(123, 4),
-        "Decimal(38,0)": DECIMAL(38, 0),
-        # Invalid decimal format
-        "Decimal(38)": str,
-        "boolean": bool,
-        "SomeRandomNotExistingType": str,
-        "bytea": bytes,
-    }
-    array_types = {f"array({k})": ARRAY(v) for k, v in base_types.items()}
-    nullable_arrays = {f"{k} null": v for k, v in array_types.items()}
-    nested_arrays = {f"array({k})": ARRAY(v) for k, v in array_types.items()}
-    return {**base_types, **array_types, **nullable_arrays, **nested_arrays}
