+++ tests/unit/common/test_typing_format.py
